Move FtResult in its own header
